refactor: move template editor context
