+++ src/core/contexts/TemplateEditorContext.tsx
